[android] Declare touchscreen as unneeded
We do not require touchscreen (though we can use it). Declare it as not
required in the manifest (it defaults to required as per documentation).

This is mostly cosmetic at this point, since it mostly affects Play
Store filtering, but it does not hurt to have the manifest correct.

Conflicts:
	tools/android/packaging/xbmc/AndroidManifest.xml