Inertial scrolling: Remove jerk
Inertial scrolling code included the decrease of the deacceleration
during the scrolling, resulting in non-zero jerk - but the math used
in the calculations does not add up. Also the conditions for checking
the time during which to apply the decrease are mixed up.
So go back to plain deacceleration with zero jerk, it at least works
reliably. If this is still needed, it must be done differently with
correct math.

Signed-off-by: Philipp Kerling <pkerling@casix.org>